server/pyproject.toml
[tool.poetry]
name = "baid-server"
version = "0.1.0"
description = ""
authors = ["Sriram K K <sriram@beskar.tech>"]

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
pydantic = {extras = ["email"], version = "^2.11.4"}
pydantic-settings = "^2.9.1"
structlog = "^25.3.0"
python-dotenv = "^1.0.0"
google-cloud-aiplatform = {extras = ["agent_engines", "adk"], version = "1.92.0"}
google-cloud-storage = "^2.18.0"
httpx = "^0.28.1"
python-jose = "^3.3.0"
aiofiles = "^23.2.1"
asyncpg = "0.30.0"
google-cloud-secret-manager = "^2.22.0"
psycopg2-binary = "^2.9.9"
setuptools = "^80.1.0"
wheel = "^0.43.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
tree-sitter = "0.19.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
